// types/note.ts
// export interface Note {
//   id: string;
//   title: string;
//   content: string;
//   tag: string;
// }


// types/note.ts
export interface Note {
  id: string;
  title: string;
  content: string;
  tag: string;
  createdAt: string; // додано
  updatedAt: string; // додано
}


// export type NoteTag = "Todo" | "Work" | "Personal" | "Meeting" | "Shopping";

// export interface Note {
//   id: string;
//   title: string;
//   content: string;
//   tag: NoteTag;
//   createdAt: string;
//   updatedAt: string;
// }